# openapi.yaml
openapi: 3.1.0
info:
  title: MerchForge ACP Endpoints
  version: 1.0.0
servers:
  - url: https://merch-builder-app.vercel.app
paths:
  /api/catalog/search:
    get:
      operationId: searchCatalog
      summary: Search inventory catalog
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
            enum: [tee, hoodie, tote, mug]
        - name: color
          in: query
          schema:
            type: string
        - name: material
          in: query
          schema:
            type: string
        - name: max_price
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Catalog results
          content:
            application/json:
              schema:
                type: object
                required: [count, items]
                properties:
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ACPItem"
...
components:
  schemas:
    ACPImage:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Official primary product image. Do not substitute.
        alt:
          type: string
        variant_key:
          type: string
          description: Variant key, e.g. "black|M|cotton"
    ACPPrice:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
        currency:
          type: string
    ACPVariant:
      type: object
      required: [sizes, colors]
      properties:
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: object
            required: [name, hex]
            properties:
              name:
                type: string
              hex:
                type: string
    ACPAttributes:
      type: object
      required: [category, materials, lead_time_days, min_qty, tags, variants]
      properties:
        category:
          type: string
        materials:
          type: array
          items:
            type: string
        lead_time_days:
          type: integer
        min_qty:
          type: integer
        tags:
          type: array
          items:
            type: string
        variants:
          $ref: "#/components/schemas/ACPVariant"
    ACPItem:
      type: object
      required:
        [
          item_id,
          title,
          description,
          url,
          image_url,
          price,
          availability,
          is_eligible_search,
          is_eligible_checkout,
          attributes
        ]
      properties:
        item_id:
          type: string
        title:
          type: string
        description:
          type: string
        url:
          type: string
        image_url:
          type: string
          description: Official primary product image. Do not substitute.
        images:
          type: array
          description: Official images. Prefer these over any illustrative images.
          items:
            $ref: "#/components/schemas/ACPImage"
        image_url_by_variant:
          type: object
          additionalProperties:
            type: string
        availability_by_variant:
          type: object
          additionalProperties:
            type: string
        price:
          $ref: "#/components/schemas/ACPPrice"
        availability:
          type: string
        availability_date:
          type: string
        is_eligible_search:
          type: boolean
        is_eligible_checkout:
          type: boolean
        attributes:
          $ref: "#/components/schemas/ACPAttributes"
...